requirements.txt
aiohttp==3.6.2
aliexpress-page-parser==0.0.9
amazon-mws==0.2.3
appdirs==1.4.3
asn1crypto==1.2.0
astroid==2.3.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bit==0.6.0
boto==2.49.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
clarifai==2.6.2
Click==7.0
coincurve==13.0.0
colorama==0.3.9
coloredlogs==10.0
configparser==4.0.2
cryptography==2.8
cssselect==1.1.0
cursor==1.2.0
cycler==0.10.0
decorator==4.0.11
docopt==0.6.2
docutils==0.15.2
dukpy==0.2.2
EasyProcess==0.2.7
ebaysdk==2.1.5
emoji==0.5.4
entrypoints==0.3
enum34==1.1.6
fake-useragent==0.1.11
feedparser==5.2.1
filetype==1.0.5
flake8==3.7.9
Flask==1.0.2
Flask-Login==0.4.1
freeze==1.0.10
fs==2.4.11
future==0.18.2
googleapis-common-protos==1.6.0
grpcio==1.25.0
halo==0.0.23
humanfriendly==4.18
idna==2.8
imageio==2.1.2
importlib-metadata==0.23
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instaloader==4.2.10
instalooter==2.4.0
instapy==0.6.6
instapy-cli==0.0.12
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
jupyter-core==4.6.1
Kivy==1.10.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
log-symbols==0.0.12
lorem==0.1.1
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.0.3
maxminddb==1.5.1
mccabe==0.6.1
MeaningCloud-python==1.1.1
more-itertools==7.2.0
MouseInfo==0.1.2
moviepy==0.2.3.2
multidict==4.5.2
names==0.3.0
nbformat==4.4.0
numpy==1.16.3
oauthlib==3.1.0
ordereddict==1.1
parsel==1.5.2
phonenumbers==8.10.12
Pillow==6.2.1
pipreqs==0.4.9
plotly==3.9.0
plyer==1.4.2
praw==6.5.1
prawcore==1.0.1
protobuf==3.10.0
psaw==0.0.8
psutil==5.6.2
psycopg2-binary==2.8.2
PyAutoGUI==0.9.48
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyGetWindow==0.0.8
Pygments==2.4.2
pylint==2.4.3
PyMsgBox==1.0.7
pync==2.0.3
pyobjc==6.1
pyobjc-core==6.1
pyobjc-framework-Accounts==6.1
pyobjc-framework-AddressBook==6.1
pyobjc-framework-AdSupport==6.1
pyobjc-framework-AppleScriptKit==6.1
pyobjc-framework-AppleScriptObjC==6.1
pyobjc-framework-ApplicationServices==6.1
pyobjc-framework-Automator==6.1
pyobjc-framework-AVFoundation==6.1
pyobjc-framework-AVKit==6.1
pyobjc-framework-BusinessChat==6.1
pyobjc-framework-CalendarStore==6.1
pyobjc-framework-CFNetwork==6.1
pyobjc-framework-CloudKit==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-Collaboration==6.1
pyobjc-framework-ColorSync==6.1
pyobjc-framework-Contacts==6.1
pyobjc-framework-ContactsUI==6.1
pyobjc-framework-CoreAudio==6.1
pyobjc-framework-CoreAudioKit==6.1
pyobjc-framework-CoreBluetooth==6.1
pyobjc-framework-CoreData==6.1
pyobjc-framework-CoreLocation==6.1
pyobjc-framework-CoreMedia==6.1
pyobjc-framework-CoreMediaIO==6.1
pyobjc-framework-CoreML==6.1
pyobjc-framework-CoreServices==6.1
pyobjc-framework-CoreSpotlight==6.1
pyobjc-framework-CoreText==6.1
pyobjc-framework-CoreWLAN==6.1
pyobjc-framework-CryptoTokenKit==6.1
pyobjc-framework-DictionaryServices==6.1
pyobjc-framework-DiscRecording==6.1
pyobjc-framework-DiscRecordingUI==6.1
pyobjc-framework-DiskArbitration==6.1
pyobjc-framework-DVDPlayback==6.1
pyobjc-framework-EventKit==6.1
pyobjc-framework-ExceptionHandling==6.1
pyobjc-framework-ExternalAccessory==6.1
pyobjc-framework-FinderSync==6.1
pyobjc-framework-FSEvents==6.1
pyobjc-framework-GameCenter==6.1
pyobjc-framework-GameController==6.1
pyobjc-framework-GameKit==6.1
pyobjc-framework-GameplayKit==6.1
pyobjc-framework-ImageCaptureCore==6.1
pyobjc-framework-IMServicePlugIn==6.1
pyobjc-framework-InputMethodKit==6.1
pyobjc-framework-InstallerPlugins==6.1
pyobjc-framework-InstantMessage==6.1
pyobjc-framework-Intents==6.1
pyobjc-framework-IOSurface==6.1
pyobjc-framework-iTunesLibrary==6.1
pyobjc-framework-LatentSemanticMapping==6.1
pyobjc-framework-LaunchServices==6.1
pyobjc-framework-libdispatch==6.1
pyobjc-framework-LocalAuthentication==6.1
pyobjc-framework-MapKit==6.1
pyobjc-framework-MediaAccessibility==6.1
pyobjc-framework-MediaLibrary==6.1
pyobjc-framework-MediaPlayer==6.1
pyobjc-framework-MediaToolbox==6.1
pyobjc-framework-MetalKit==6.1
pyobjc-framework-ModelIO==6.1
pyobjc-framework-MultipeerConnectivity==6.1
pyobjc-framework-NaturalLanguage==6.1
pyobjc-framework-NetFS==6.1
pyobjc-framework-Network==6.1
pyobjc-framework-NetworkExtension==6.1
pyobjc-framework-NotificationCenter==6.1
pyobjc-framework-OpenDirectory==6.1
pyobjc-framework-OSAKit==6.1
pyobjc-framework-Photos==6.1
pyobjc-framework-PhotosUI==6.1
pyobjc-framework-PreferencePanes==6.1
pyobjc-framework-PubSub==6.1
pyobjc-framework-QTKit==6.1
pyobjc-framework-Quartz==6.1
pyobjc-framework-SafariServices==6.1
pyobjc-framework-SceneKit==6.1
pyobjc-framework-ScreenSaver==6.1
pyobjc-framework-ScriptingBridge==6.1
pyobjc-framework-SearchKit==6.1
pyobjc-framework-Security==6.1
pyobjc-framework-SecurityFoundation==6.1
pyobjc-framework-SecurityInterface==6.1
pyobjc-framework-ServiceManagement==6.1
pyobjc-framework-Social==6.1
pyobjc-framework-SpriteKit==6.1
pyobjc-framework-StoreKit==6.1
pyobjc-framework-SyncServices==6.1
pyobjc-framework-SystemConfiguration==6.1
pyobjc-framework-UserNotifications==6.1
pyobjc-framework-VideoSubscriberAccount==6.1
pyobjc-framework-VideoToolbox==6.1
pyobjc-framework-Vision==6.1
pyobjc-framework-WebKit==6.1
pyparsing==2.4.4
pyperclip==1.7.0
PyRect==0.1.4
pyrsistent==0.15.5
PyScreeze==0.1.25
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==12.2.0
PyTweening==1.0.3
pytz==2019.3
PyVirtualDisplay==0.2.1
pyxdg==0.26
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
scipy==1.2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.5
spinners==0.0.23
tenacity==5.1.5
termcolor==1.1.0
tornado==6.0.3
tqdm==4.11.2
traitlets==4.3.3
transmissionrpc==0.11
tweepy==3.8.0
typed-ast==1.4.0
update-checker==0.16
urllib3==1.25.6
verboselogs==1.7
w3lib==1.21.0
wapy==1.0.0
webdriverdownloader==1.0.0.1
websocket-client==0.57.0
Werkzeug==0.16.0
WooCommerce==2.1.1
wordpress-api==1.2.9
wrapt==1.11.2
xmltodict==0.12.0
yapf==0.28.0
yarg==0.1.9
yarl==1.1.0
yaspin==0.14.3
zipp==0.6.0
